Changed return value

+++ DBMS_BookStore/DAO/KhachHangDAO.cs
@@ -76,12 +76,12 @@ namespace DBMS_BookStore.DAO
             return (int)DBConnection.ExecuteScalar(cmd);
         }
 
-        public double GetDiscountValue(string MaKH)
+        public float GetDiscountValue(string MaKH)
         {
             string query = "SELECT dbo.FUNC_GetDiscount(@param)";
             SqlCommand cmd = new SqlCommand(query);
             cmd.Parameters.AddWithValue("@param", MaKH);
-            return (double)DBConnection.ExecuteScalar(cmd);
+            return (float)DBConnection.ExecuteScalar(cmd);
         }
     }
 }
